Add a guard page after each huge allocation
This will catch overflows in huge allocations, for the low low price of one page
(4k usually) per chunk (2M). An easy followup would be to add a guard page
before the chunk, to catch underflows.

This technique is used in several other allocators, like:

- https://chromium.googlesource.com/chromium/src/+/master/base/allocator/partition_allocator/PartitionAlloc.md
- https://source.android.com/docs/security/test/scudo/
- https://github.com/struct/isoalloc/
- https://github.com/microsoft/mimalloc/ (optionally)
- https://musl.libc.org/
- https://github.com/GrapheneOS/hardened_malloc/
- https://github.com/microsoft/snmalloc/

The implementation is inspired from the fiber's one.